requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.1
python-resize-image==1.1.19
pytz==2019.3
redis==3.4.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
